backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
django-filter>=24.2
django-allauth>=0.63
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.4
drf-yasg>=1.21.7

# импорт Excel
pandas>=2.2.2
openpyxl>=3.1.5